move core parser to library
